requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
jinja2>=3.1.4,<4.0.0
sqlalchemy>=2.0.30,<3.0.0
alembic>=1.13.1,<2.0.0
psycopg[binary]>=3.1.19,<4.0.0
httpx>=0.27.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
apscheduler>=3.10.4,<4.0.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
